fix: prevent uint32 overflow in DD-mode filter buffer indexing
bfSize and bfCount are uint32, and several call sites computed byte
offsets into sd.buffer as raw uint32 products (bfSize*i). For large
DD-mode digests (e.g. a ~1 GiB input with a 64-byte block size,
yielding bfCount ~= 2^24) this product overflows 32 bits, producing
either an inverted slice that panics or a wrapped-but-in-bounds index
that silently corrupts the digest.

Promote all such offset arithmetic to uint64 before slicing:

- sdhash.go: computeHamming and the DD branch of String()
- generate.go: the post-generation buffer trim in both the
  single-chunk and multi-chunk code paths
- score.go: sdbfMaxScore's reference/target filter offsets, reachable
  via Similarity() on large DD-mode digests (same bug class, not
  explicitly in the original report but sharing the identical
  pattern)

No behavior change for inputs within the previous safe range;
verified against the existing test suite and against a targeted
overflow repro (bfSize=256, i=2^24-1) that reproduces the exact
inverted-slice condition on the old code and confirms correct,
monotonic bounds on the fixed code.

Fixes #62

+++ score.go
@@ -65,16 +65,16 @@ func sdbfScore(sdbf1 *sdbf, sdbf2 *sdbf) int {
 // filter had enough elements to score against, or a value in [0.0, 1.0] otherwise.
 func sdbfMaxScore(refSdbf *sdbf, refIndex uint32, targetSdbf *sdbf) float64 {
 	var maxScore float64 = -1
-	bfSize := refSdbf.bfSize
+	bfSize := uint64(refSdbf.bfSize)
 
 	s1 := refSdbf.elemCount(refIndex)
 	if s1 < minElemCount {
 		return 0
 	}
-	bf1 := refSdbf.buffer[refIndex*bfSize:]
+	bf1 := refSdbf.buffer[uint64(refIndex)*bfSize:]
 	e1Cnt := refSdbf.hamming[refIndex]
 	for i := range targetSdbf.bfCount {
-		bf2 := targetSdbf.buffer[i*bfSize:]
+		bf2 := targetSdbf.buffer[uint64(i)*bfSize:]
 		s2 := targetSdbf.elemCount(i)
 		if s2 < minElemCount {
 			continue
